+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- C bindings to Cython bindings
+
 ## [0.2.0] - 2016-07-05
 ### Added
 - window_sz2 and lookahead_sz2 parameters for encode/decode functions.
@@ -11,4 +15,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Performance of encoding/decoding large files
 - Compilation for python 3 (still not fully working)
 
+[Unreleased]: https://github.com/johan-sports/pyheatshrink/compare/0.2.0..HEAD
 [0.2.0]: https://github.com/johan-sports/pyheatshrink/compare/0.1.2...0.2.0